requirements.txt
requests==2.31.0
python-dotenv==1.0.1
anthropic==0.40.0
google-genai>=1.16.0
pytest==8.3.3
pytest-mock==3.14.0
